Suppression des warnings

+++ c_grid.cpp
@@ -336,10 +336,10 @@ std::vector<gf::Vector2u> Grid::getDestinations(gf::Vector2u coords) {
     {
       // Le scout peut bouger d'autant qu'il veut en ligne, sans passer au dessus des autres pièces
       // Parcours des cases autour
-      int x = coords.x - 1;
+      unsigned int x = (unsigned)(coords.x - 1);
       while(x >= 0) {
         if(grid[x][coords.y].rank == Rank::Empty || grid[x][coords.y].side == Side::Blue) {
-          destinations.push_back({x, coords.y});
+          destinations.push_back({x, (unsigned)coords.y});
         }
         
         if(grid[x][coords.y].rank != Rank::Empty || grid[x][coords.y].side == Side::Blue) {
@@ -352,7 +352,7 @@ std::vector<gf::Vector2u> Grid::getDestinations(gf::Vector2u coords) {
       x = coords.x + 1;
       while(x < GridSize) {
         if(grid[x][coords.y].rank == Rank::Empty || grid[x][coords.y].side == Side::Blue) {
-          destinations.push_back({x, coords.y});
+          destinations.push_back({x, (unsigned)coords.y});
         }
         
         if(grid[x][coords.y].rank != Rank::Empty || grid[x][coords.y].side == Side::Blue) {
@@ -362,10 +362,10 @@ std::vector<gf::Vector2u> Grid::getDestinations(gf::Vector2u coords) {
         x++;
       }
 
-      int y = coords.y - 1;
+      unsigned int y = (unsigned)(coords.y - 1);
       while(y >= 0) {
         if(grid[coords.x][y].rank == Rank::Empty || grid[coords.x][y].side == Side::Blue) {
-          destinations.push_back({coords.x, y});
+          destinations.push_back({(unsigned)coords.x, y});
         }
         
         if(grid[coords.x][y].rank != Rank::Empty || grid[coords.x][y].side == Side::Blue) {
@@ -378,7 +378,7 @@ std::vector<gf::Vector2u> Grid::getDestinations(gf::Vector2u coords) {
       y = coords.y + 1;
       while(y < GridSize) {
         if(grid[coords.x][y].rank == Rank::Empty || grid[coords.x][y].side == Side::Blue) {
-          destinations.push_back({coords.x, y});
+          destinations.push_back({(unsigned)coords.x, y});
         }
         
         if(grid[coords.x][y].rank != Rank::Empty || grid[coords.x][y].side == Side::Blue) {
